requirements.txt
Flask==2.3.2
requests==2.31.0
python-dotenv==1.0.0
pytest==7.3.1
flake8==6.0.0
gunicorn==20.1.0
